Worker erstellt, AffiliationWorker implementiert, bindings geupdated

+++ MethodDemos/src/demos/Demos.java
@@ -1,11 +1,19 @@
 package demos;
+
 import java.io.File;
 import java.io.IOException;
 import java.util.Arrays;
 import java.util.List;
 import java.util.stream.Collectors;
 
+import javax.xml.bind.JAXBException;
+
 import org.apache.commons.io.FileUtils;
+
+import mapping.cermine.CermineMapper;
+import mapping.grobid.GROBIDMapper;
+import mapping.parscit.ParsCitMapper;
+import mapping.pdfx.PDFXMapper;
 
 /**
  * Runs all Demos with Ground Truth files
@@ -15,41 +23,92 @@ public class Demos
 {
 	private static final boolean USE_SPECIFIC_OUTPUTDIR = false;
 
-	private static File inputDir = new File("D:/output/GroundTruth");
-	private static File inputDirTxt = new File("D:/output/Cermine");
+	public static File inputDir = new File("D:/output/GroundTruth");
+	// public static File inputDirTxt = new File("D:/output/Cermine");
 
 	static File cermineOutputDir = new File("D:/output/Cermine");
 	static File grobIdOutputDir = new File("D:/output/Grobid");
 	static File parsCitOutputDir = new File("D:/output/ParsCit");
 	static File pdfxOutputDir = new File("D:/output/PDFX");
+	static File[] outputDirs =
+	{cermineOutputDir, grobIdOutputDir, parsCitOutputDir, pdfxOutputDir};
 
 	static File allOutputDir = new File("D:/output/all");
 
-	public static void main(String[] args) throws IOException
+	public static void main(String[] args) throws Exception
 	{
-		List<File> groundTruthFiles = getAllGroundTruthFiles().subList(0, 1);
+		executeDemos();
+	}
 
-		FileUtils.cleanDirectory(allOutputDir);
-		FileUtils.cleanDirectory(cermineOutputDir);
-		FileUtils.cleanDirectory(grobIdOutputDir);
-		FileUtils.cleanDirectory(parsCitOutputDir);
-		FileUtils.cleanDirectory(pdfxOutputDir);
+	private static void executeDemos() throws IOException, JAXBException
+	{
+		boolean runDemos = false;
+		boolean runCermineDemo = false;
+		boolean runGrobidDemo = false;
+		boolean runParsCitDemo = true;
+		boolean runPdfxDemo = false;
+		boolean runCermineMapper = true;
+		boolean runGrobidMapper = true;
+		boolean runParsCitMapper = true;
+		boolean runPdfxMapper = true;
+		List<File> groundTruthFiles = getAllGroundTruthFiles().subList(0, 10);
+		List<File> groundTruthFilesOmnipage = getAllGroundTruthFilesAsOmnipage().subList(0, 10);
 
 		if(!USE_SPECIFIC_OUTPUTDIR)
 		{
+			if(runDemos)
+			{
+				FileUtils.cleanDirectory(allOutputDir);
+			}
+			deleteResultAndErrorFiles(allOutputDir);
+
 			cermineOutputDir = allOutputDir;
 			grobIdOutputDir = allOutputDir;
 			parsCitOutputDir = allOutputDir;
 			pdfxOutputDir = allOutputDir;
-
-			inputDirTxt = allOutputDir;
+		}
+		else
+		{
+			if(runDemos)
+			{
+				if(runCermineDemo) FileUtils.cleanDirectory(cermineOutputDir);
+				if(runGrobidDemo) FileUtils.cleanDirectory(grobIdOutputDir);
+				if(runParsCitDemo) FileUtils.cleanDirectory(parsCitOutputDir);
+				if(runPdfxDemo) FileUtils.cleanDirectory(pdfxOutputDir);
+			}
+			if(runCermineMapper) deleteResultAndErrorFiles(cermineOutputDir);
+			if(runGrobidMapper) deleteResultAndErrorFiles(grobIdOutputDir);
+			if(runParsCitMapper) deleteResultAndErrorFiles(parsCitOutputDir);
+			if(runPdfxMapper) deleteResultAndErrorFiles(pdfxOutputDir);
 		}
 
-		new CERMINEDemo().runDemo(groundTruthFiles, cermineOutputDir);
-		new GROBIDDemo().runDemo(groundTruthFiles, grobIdOutputDir);
-		new ParsCitDemo().runDemo(getAllGroundTruthFilesAsTxt(), parsCitOutputDir); // uses txt output from Cermine
-		new PDFXDemo().runDemo(groundTruthFiles, pdfxOutputDir);
+		if(runDemos)
+		{
+			if(runCermineDemo) new CERMINEDemo().runDemo(groundTruthFiles, cermineOutputDir);
+			if(runGrobidDemo) new GROBIDDemo().runDemo(groundTruthFiles, grobIdOutputDir);
+			if(runParsCitDemo) new ParsCitDemo().runDemo(groundTruthFilesOmnipage, parsCitOutputDir); // uses txt output from Cermine
+			if(runPdfxDemo) new PDFXDemo().runDemo(groundTruthFiles, pdfxOutputDir);
+		}
 
+		if(runCermineMapper) new CermineMapper().unmarshallFiles(cermineOutputDir);
+		if(runGrobidMapper) new GROBIDMapper().unmarshallFiles(grobIdOutputDir);
+		if(runParsCitMapper) new ParsCitMapper().unmarshallFiles(parsCitOutputDir);
+		if(runPdfxMapper) new PDFXMapper().unmarshallFiles(pdfxOutputDir);
+	}
+
+	private static void deleteResultAndErrorFiles(File directory)
+	{
+		List<File> resultOrErrorFiles = Arrays.stream(directory.listFiles()).filter(file -> file.getName().endsWith("-err.xml") || file.getName().endsWith("-xstream.xml")).collect(Collectors.toList());
+		for(File file : resultOrErrorFiles)
+		{
+			file.delete();
+		}
+	}
+
+	private static List<File> getAllGroundTruthFilesAsOmnipage()
+	{
+		List<File> groundTruthFiles = getAllFilesAlsoFromSubDirectories(inputDir, "-omnipage.xml");
+		return groundTruthFiles;
 	}
 
 	public static List<File> getAllGroundTruthFiles()
@@ -58,11 +117,11 @@ public class Demos
 		return groundTruthFiles;
 	}
 
-	public static List<File> getAllGroundTruthFilesAsTxt()
-	{
-		List<File> groundTruthFiles = getAllFilesFromDirectories(inputDirTxt, ".txt");
-		return groundTruthFiles;
-	}
+	// public static List<File> getAllGroundTruthFilesAsTxt()
+	// {
+	// List<File> groundTruthFiles = getAllFilesFromDirectories(inputDirTxt, ".txt");
+	// return groundTruthFiles;
+	// }
 
 	public static List<File> getAllFilesAlsoFromSubDirectories(File inputFolder, String extension)
 	{
